--- config/attack/PGD.yaml
@@ -1,18 +1,0 @@
-name: PGDAttack
-short_name: pgd_attack
-
-attack_params:
-  eps: [0.03]
-  n_steps: 10
-  eta: 0.5
-  norm: null
-
-hyperparameters_vary:
-  eps:
-    optuna_type: float
-    low: 0.01
-    high: 0.5
-    log: True
-  n_steps:
-    optuna_type: const
-    value: 50
