+++ config_examples/attack/FGSMRegDiscSmoothMaxAttack.yaml
