+++ config_examples/attack_run_config.yaml
@@ -7,7 +7,6 @@ exp_disc_name: _test
 log_clearml: False
 clearml_project: "AA_attack_run"
 author: '' # Petr, Dima, Sophie
-pretrained: False
 
 model_folder: checkpoints/${dataset.name}/${attack_model.name}/${exp_model_name}
 disc_path: results/${dataset.name}/${attack_model.name}/discriminator/fgsm_attack${exp_disc_name}
@@ -22,10 +21,10 @@ hydra:
       ++attack.attack_params.eps: 0.01,0.03
 
 defaults:
-  - dataset: Strawberry
-  - model@attack_model: TST
-  - model@disc_model_reg: TST
-  - model@disc_model_check: TST
+  - dataset: PowerCons
+  - model@attack_model: ResidualCNN
+  - model@disc_model_reg: ResidualCNN
+  - model@disc_model_check: ResidualCNN
   - attack: FGSMDisc
 
 criterion_name: BCELoss
