+++ config_examples/dataset/ProximalPhalanxOutlineCorrect.yaml
