+++ config_examples/dataset/Strawberry.yaml
