+++ config_examples/model/ResidualCNN.yaml
