+++ src/attacks/__init__.py
@@ -1,4 +1,3 @@
 from .base_attacks import *
 from .fgsm import *
 from .simba import *
-from .pgd import *
