+++ src/attacks/simba.py
@@ -1,92 +1,78 @@
 import torch
 
-from src.attacks.base_attacks import BaseIterativeAttack
-from src.attacks.procedures import BatchIterativeAttack
+from .base_attacks import BaseIterativeAttack
 from .regularizers import reg_disc, reg_neigh
 
 
-class SimBABinary(BaseIterativeAttack, BatchIterativeAttack):
-    def __init__(self, model, criterion, eps, n_steps, estimator, device="cpu", **kwargs):
+class SimBABinary(BaseIterativeAttack):
+    def __init__(self, model, criterion, eps, n_steps, device="cpu"):
         super().__init__(model, criterion, eps, n_steps, device=device)
-        BaseIterativeAttack.__init__(self, model=model, n_steps=n_steps)
-        BatchIterativeAttack.__init__(self, estimator=estimator)
-        self.criterion = criterion
-        self.criterion.reduction='none'
-        self.is_regularized = False
 
-        self.eps = eps
+    def step(self, x, y_true):
+        rand_ind = torch.randint(x.shape[1], (1,))[0]
+        mask = torch.zeros(x.shape).to(self.device)
+        mask[:, rand_ind] = mask[:, rand_ind] + 1
 
-    def get_loss(self, X: torch.Tensor, y_true: torch.Tensor) -> torch.Tensor:
-        self.model.zero_grad()
-        y_pred = self.model(X)
-        loss = self.criterion(y_pred, y_true)
-        return loss
+        x_minus = x - mask * self.eps
+        x_plus = x + mask * self.eps
 
-    def generate_changes(self, X):
-        random_indices = torch.randint(0, X.shape[1], (X.shape[0], 1, 1)).to(self.device)
+        y_pred = self.model(x)
+        y_pred_minus = self.model(x_minus)
+        y_pred_plus = self.model(x_plus)
 
-        ones = torch.ones_like(X)
-        X_minus = X.scatter_add(1, random_indices, -self.eps * ones)
-        X_plus = X.scatter_add(1, random_indices, self.eps * ones)
+        p_orig = torch.sigmoid(y_pred)
+        p_minus = torch.sigmoid(y_pred_minus)
+        p_plus = torch.sigmoid(y_pred_plus)
 
-        return X_minus, X_plus
-
-
-    def step(self, X, y_true):
-        X_minus, X_plus = self.generate_changes(X)
-
-        y_pred = self.model(X)
-        y_pred_minus = self.model(X_minus)
-        y_pred_plus = self.model(X_plus)
-
-        data_all = torch.cat([X, X_minus, X_plus], dim=-1)
-
-        losses = torch.stack([
-            self.criterion(y_pred, y_true), 
-            self.criterion(y_pred_minus, y_true), 
-            self.criterion(y_pred_plus, y_true)], dim=0)
-
-        ind = losses.argmax(0)
-        res = torch.zeros(*data_all.shape[:-1], 1).to(self.device)
-
-        for i, row in enumerate(data_all):
-            res[i, :, :] = row[:, ind[i].item()].unsqueeze(-1)
-
-        return res
-
-
-class SimBABinaryDiscReg(SimBABinary):
-    def __init__(
-        self, model, criterion, eps, n_steps, alpha, disc_models, device="cpu", use_sigmoid=False,
-    ):
-        super().__init__(model, criterion, eps, n_steps, device=device)
-        self.use_sigmoid = use_sigmoid
-        self.alpha = alpha
-        self.disc_models = disc_models
-
-    def step(self, X, y_true):
-        X_minus, X_plus = self.generate_changes(X)
-
-        y_pred = self.model(X)
-        y_pred_minus = self.model(X_minus)
-        y_pred_plus = self.model(X_plus)
-
-        data_all = torch.cat([X, X_minus, X_plus], dim=-1)
-
+        x_all = torch.cat([x, x_minus, x_plus]).view(3, *x.shape)
         losses = torch.stack(
             [
-                self.criterion(y_pred, y_true) - self.alpha * reg_disc(X, self.disc_models, self.use_sigmoid),
-                self.criterion(y_pred_minus, y_true) - self.alpha * reg_disc(X_minus, self.disc_models, self.use_sigmoid),
-                self.criterion(y_pred_plus, y_true) - self.alpha * reg_disc(X_plus, self.disc_models, self.use_sigmoid),
+                self.criterion(p_orig, y_true),
+                self.criterion(p_minus, y_true),
+                self.criterion(p_plus, y_true),
             ],
             dim=0,
         )
 
-        ind = losses.argmax(0)
-        res = torch.zeros(*data_all.shape[:-1], 1).to(self.device)
+        min_loss_indices = torch.argmin(losses, dim=0)
+        min_loss_indices = (
+            min_loss_indices.unsqueeze(-1).unsqueeze(-1).expand(-1, x.shape[-1])
+        )
 
-        for i, row in enumerate(data_all):
-            res[i, :, :] = row[:, ind[i].item()].unsqueeze(-1)
+        x_adv = torch.gather(x_all, 0, min_loss_indices).squeeze(0)
+        return x_adv
 
-        return res
 
+class SimBABinaryReg(SimBABinary):
+    def __init__(self, model, criterion, eps, n_steps, alpha, device="cpu"):
+        super().__init__(model, criterion, eps, n_steps, device=device)
+        self.alpha = alpha
+
+    def step(self, x, y_true):
+        x_adv = super().step(x, y_true)
+        reg_value = reg_neigh(x_adv, self.alpha)
+
+        loss = -reg_value
+        grad_ = torch.autograd.grad(loss, x_adv, retain_graph=True)[0]
+        x_adv = x_adv.data + grad_
+
+        return x_adv
+
+
+class SimBABinaryDiscReg(SimBABinary):
+    def __init__(
+        self, model, criterion, eps, n_steps, alpha, disc_models, device="cpu"
+    ):
+        super().__init__(model, criterion, eps, n_steps, device=device)
+        self.alpha = alpha
+        self.disc_models = disc_models
+
+    def step(self, x, y_true):
+        x_adv = super().step(x, y_true)
+        reg_value = reg_disc(x_adv, self.alpha, self.disc_models)
+
+        loss = -reg_value
+        grad_ = torch.autograd.grad(loss, x_adv, retain_graph=True)[0]
+        x_adv = x_adv.data + grad_
+
+        return x_adv
