+++ train_discriminator.py
@@ -13,12 +13,12 @@ from src.config import get_criterion, get_disc_list, get_model
 from src.data import MyDataset, load_data, transform_data
 from src.estimation.estimators import AttackEstimator
 from src.training.train import DiscTrainer
-from src.utils import fix_seed, save_config, save_compiled_config,weights_from_clearml_by_name
+from src.utils import fix_seed, save_config, save_compiled_config
 
 warnings.filterwarnings("ignore")
 
 CONFIG_NAME = "train_disc_config"
-CONFIG_PATH = "config/my_configs"
+CONFIG_PATH = "config"
 
 torch.cuda.empty_cache()
 @hydra.main(config_path=CONFIG_PATH, config_name=CONFIG_NAME, version_base=None)
@@ -82,18 +82,11 @@ def main(cfg: DictConfig):
 
     device = torch.device(cfg["device"] if torch.cuda.is_available() else "cpu")
 
-    if cfg['pretrained']:
-        project_name = 'AA_train_classifiers'
-        task_name = f"model_{cfg['model']['name']}_{cfg['model_id_attack']}_{cfg['dataset']['name']}"
 
-        path = weights_from_clearml_by_name(project_name=project_name, task_name=task_name)
-        attack_model_path = os.path.join(path)
-    else:
-
-        attack_model_path = os.path.join(
-            cfg["model_folder"],
-            f"model_{cfg['model']['name']}_{cfg['model_id_attack']}_{cfg['dataset']['name']}.pt",
-        )
+    attack_model_path = os.path.join(
+        cfg["model_folder"],
+        f"model_{cfg['model']['name']}_{cfg['model_id_attack']}_{cfg['dataset']['name']}.pt",
+    )
 
     attack_model = get_model(
         cfg["attack_model"]["name"],
